Added support for custom health checks to `JournalOptionsBuilder`
